+++ test-app/app/src/main/assets/app/mainpage.js
@@ -54,3 +54,8 @@ require("./tests/testInterfaceImplementation");
 require("./tests/testRuntimeImplementedAPIs");
 require("./tests/testsInstanceOfOperator");
 require("./tests/testReleaseNativeCounterpart");
+require("./tests/kotlin/companions/testCompanionObjectsSupport");
+require("./tests/kotlin/properties/testPropertiesSupport");
+require("./tests/kotlin/delegation/testDelegationSupport");
+require("./tests/kotlin/objects/testObjectsSupport");
+require("./tests/kotlin/functions/testTopLevelFunctionsSupport");