+++ test-app/app/src/main/java/com/tns/tests/kotlin/companions/ChildKotlinClass.kt
@@ -1,0 +1,5 @@
+package com.tns.tests.kotlin.companions
+
+class ChildKotlinClass : BaseKotlinClassWithCompanion() {
+    fun getStringFromCompanion() = BaseKotlinClassWithCompanion.getString()
+}