+++ test-app/app/src/main/java/com/tns/tests/kotlin/delegation/StringGetterImpl.kt
@@ -1,0 +1,5 @@
+package com.tns.tests.kotlin.delegation
+
+class StringGetterImpl: StringGetter {
+    override fun getString(): String = "some string"
+}