+++ test-app/build-tools/android-metadata-generator/src/src/com/telerik/metadata/bcl/ClassInfo.java
@@ -1,11 +1,16 @@
 package com.telerik.metadata.bcl;
 
 import com.telerik.metadata.ClassUtil;
-import com.telerik.metadata.desc.MetadataInfoAnnotationDescriptor;
 import com.telerik.metadata.desc.ClassDescriptor;
 import com.telerik.metadata.desc.FieldDescriptor;
+import com.telerik.metadata.desc.KotlinClassMetadataAnnotation;
+import com.telerik.metadata.desc.MetadataInfoAnnotationDescriptor;
 import com.telerik.metadata.desc.MethodDescriptor;
+import com.telerik.metadata.desc.PropertyDescriptor;
+import com.telerik.metadata.kotlin.classes.KotlinClassMetadataParser;
+import com.telerik.metadata.kotlin.classes.impl.KotlinClassMetadataParserImpl;
 
+import org.apache.bcel.classfile.AnnotationEntry;
 import org.apache.bcel.classfile.Attribute;
 import org.apache.bcel.classfile.ConstantClass;
 import org.apache.bcel.classfile.ConstantPool;
@@ -15,6 +20,15 @@ import org.apache.bcel.classfile.InnerClass;
 import org.apache.bcel.classfile.InnerClasses;
 import org.apache.bcel.classfile.JavaClass;
 import org.apache.bcel.classfile.Method;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+
+import kotlinx.metadata.KmProperty;
+import kotlinx.metadata.jvm.JvmExtensionsKt;
+import kotlinx.metadata.jvm.JvmMethodSignature;
 
 public class ClassInfo implements ClassDescriptor {
     private final JavaClass clazz;
@@ -66,11 +80,11 @@ public class ClassInfo implements ClassDescriptor {
 
     private String getClassName(int classIndex) {
         ConstantPool constantPool = clazz.getConstantPool();
-        ConstantClass innerClassNameIndex = (ConstantClass)constantPool.getConstant(classIndex);
+        ConstantClass innerClassNameIndex = (ConstantClass) constantPool.getConstant(classIndex);
         if (innerClassNameIndex == null) {
             return null;
         }
-        ConstantUtf8 className = (ConstantUtf8)constantPool.getConstant(innerClassNameIndex.getNameIndex());
+        ConstantUtf8 className = (ConstantUtf8) constantPool.getConstant(innerClassNameIndex.getNameIndex());
         if (className == null) {
             return null;
         }
@@ -111,17 +125,57 @@ public class ClassInfo implements ClassDescriptor {
     public MethodDescriptor[] getMethods() {
         Method[] ms = clazz.getMethods();
         MethodDescriptor[] methods = new MethodDescriptor[ms.length];
-        for (int i=0; i<methods.length; i++) {
+        for (int i = 0; i < methods.length; i++) {
             methods[i] = new MethodInfo(ms[i]);
         }
         return methods;
     }
 
     @Override
+    public PropertyDescriptor[] getProperties() {
+        List<PropertyDescriptor> propertyDescriptors = new ArrayList<PropertyDescriptor>();
+        KotlinClassMetadataParser kotlinClassMetadataParser = new KotlinClassMetadataParserImpl();
+        List<KmProperty> kotlinProperties = kotlinClassMetadataParser.getKotlinProperties(this);
+        for (KmProperty kmProperty : kotlinProperties) {
+            String propertyName = kmProperty.getName();
+
+            MethodDescriptor getter = null;
+            JvmMethodSignature getterSignature = JvmExtensionsKt.getGetterSignature(kmProperty);
+            if (getterSignature != null) {
+                getter = getMethodDescriptorWithSignature(getterSignature.getName(), getterSignature.getDesc());
+            }
+
+            MethodDescriptor setter = null;
+            JvmMethodSignature setterSignature = JvmExtensionsKt.getSetterSignature(kmProperty);
+            if (setterSignature != null) {
+                setter = getMethodDescriptorWithSignature(setterSignature.getName(), setterSignature.getDesc());
+            }
+
+            propertyDescriptors.add(new PropertyInfo(propertyName, getter, setter));
+
+        }
+
+        return propertyDescriptors.toArray(new PropertyDescriptor[0]);
+    }
+
+    private MethodDescriptor getMethodDescriptorWithSignature(String name, String signature) {
+        for (Method method : clazz.getMethods()) {
+            String methodSignature = method.getSignature();
+            String methodName = method.getName();
+
+            if (methodSignature.equals(signature) && methodName.equals(name)) {
+                return new MethodInfo(method);
+            }
+        }
+
+        return null;
+    }
+
+    @Override
     public FieldDescriptor[] getFields() {
         Field[] fs = clazz.getFields();
         FieldDescriptor[] fields = new FieldDescriptor[fs.length];
-        for (int i=0; i<fields.length; i++) {
+        for (int i = 0; i < fields.length; i++) {
             fields[i] = new FieldInfo(fs[i]);
         }
         return fields;
@@ -130,6 +184,22 @@ public class ClassInfo implements ClassDescriptor {
     @Override
     public MetadataInfoAnnotationDescriptor getMetadataInfoAnnotation() {
         return null;
+    }
+
+    @Override
+    public Optional<KotlinClassMetadataAnnotation> getKotlinClassMetadataAnnotation() {
+        AnnotationEntry[] annotationEntries = clazz.getAnnotationEntries();
+        if (annotationEntries != null) {
+            for (AnnotationEntry annotationEntry : annotationEntries) {
+                String annotationType = annotationEntry.getAnnotationType();
+                if ("Lkotlin/Metadata;".equals(annotationType)) {
+                    KotlinClassMetadataAnnotation kotlinClassMetadataAnnotation = new KotlinBytecodeClassMetadataAnnotation(annotationEntry);
+                    return Optional.of(kotlinClassMetadataAnnotation);
+                }
+            }
+        }
+
+        return Optional.empty();
     }
 
     @Override