+++ test-app/build-tools/android-metadata-generator/src/src/com/telerik/metadata/desc/ClassDescriptor.java
@@ -1,5 +1,7 @@
 package com.telerik.metadata.desc;
 
+import java.util.Optional;
+
 public interface ClassDescriptor {
     boolean isClass();
     boolean isInterface();
@@ -11,9 +13,13 @@ public interface ClassDescriptor {
 
     MethodDescriptor[] getMethods();
 
+    PropertyDescriptor[] getProperties();
+
     FieldDescriptor[] getFields();
 
     MetadataInfoAnnotationDescriptor getMetadataInfoAnnotation();
+
+    Optional<KotlinClassMetadataAnnotation> getKotlinClassMetadataAnnotation();
 
     String[] getInterfaceNames();