+++ test-app/build-tools/android-metadata-generator/src/src/com/telerik/metadata/dx/ClassInfo.java
@@ -7,15 +7,24 @@ import com.android.dex.Dex;
 import com.android.dx.rop.code.AccessFlags;
 import com.android.dx.rop.cst.CstType;
 import com.android.dx.rop.type.Type;
-import com.telerik.metadata.desc.MetadataInfoAnnotationDescriptor;
 import com.telerik.metadata.desc.ClassDescriptor;
 import com.telerik.metadata.desc.FieldDescriptor;
+import com.telerik.metadata.desc.KotlinClassMetadataAnnotation;
+import com.telerik.metadata.desc.MetadataInfoAnnotationDescriptor;
 import com.telerik.metadata.desc.MethodDescriptor;
+import com.telerik.metadata.desc.PropertyDescriptor;
+import com.telerik.metadata.kotlin.classes.KotlinClassMetadataParser;
+import com.telerik.metadata.kotlin.classes.impl.KotlinClassMetadataParserImpl;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Optional;
+
+import kotlinx.metadata.KmProperty;
+import kotlinx.metadata.jvm.JvmExtensionsKt;
+import kotlinx.metadata.jvm.JvmMethodSignature;
 
 public class ClassInfo implements ClassDescriptor {
     private final DexFile dexFile;
@@ -72,7 +81,7 @@ public class ClassInfo implements ClassDescriptor {
             ClassData data = dex.readClassData(classDef);
             ClassData.Method[] allMethods = data.allMethods();
             methods = new MethodDescriptor[allMethods.length];
-            for (int i=0; i<allMethods.length; i++) {
+            for (int i = 0; i < allMethods.length; i++) {
                 methods[i] = new MethodInfo(dexFile, allMethods[i]);
             }
         } else {
@@ -84,6 +93,44 @@ public class ClassInfo implements ClassDescriptor {
     }
 
     @Override
+    public PropertyDescriptor[] getProperties() {
+        List<PropertyDescriptor> propertyDescriptors = new ArrayList<PropertyDescriptor>();
+        KotlinClassMetadataParser kotlinClassMetadataParser = new KotlinClassMetadataParserImpl();
+        List<KmProperty> kotlinProperties = kotlinClassMetadataParser.getKotlinProperties(this);
+        for (KmProperty kmProperty : kotlinProperties) {
+            String propertyName = kmProperty.getName();
+
+            MethodDescriptor getter = null;
+            JvmMethodSignature getterSignature = JvmExtensionsKt.getGetterSignature(kmProperty);
+            if (getterSignature != null) {
+                getter = getMethodDescriptorWithSignature(getterSignature.getName(), getterSignature.getDesc());
+            }
+
+            MethodDescriptor setter = null;
+            JvmMethodSignature setterSignature = JvmExtensionsKt.getSetterSignature(kmProperty);
+            if (setterSignature != null) {
+                setter = getMethodDescriptorWithSignature(setterSignature.getName(), setterSignature.getDesc());
+            }
+
+            propertyDescriptors.add(new PropertyInfo(propertyName, getter, setter));
+
+        }
+
+        return propertyDescriptors.toArray(new PropertyDescriptor[0]);
+
+    }
+
+    private MethodDescriptor getMethodDescriptorWithSignature(String name, String signature) {
+        Optional<MethodDescriptor> method = Arrays
+                .stream(getMethods())
+                .filter(m -> m.getName().equals(name)
+                        && m.getSignature().equals(signature))
+                .findFirst();
+
+        return method.orElse(null);
+    }
+
+    @Override
     public FieldDescriptor[] getFields() {
         FieldDescriptor[] fields;
         Dex dex = dexFile.getDex();
@@ -92,7 +139,7 @@ public class ClassInfo implements ClassDescriptor {
             ClassData data = dex.readClassData(classDef);
             ClassData.Field[] allFields = data.allFields();
             fields = new FieldDescriptor[allFields.length];
-            for (int i=0; i<fields.length; i++) {
+            for (int i = 0; i < fields.length; i++) {
                 fields[i] = new FieldInfo(dexFile, allFields[i]);
             }
         } else {
@@ -104,9 +151,18 @@ public class ClassInfo implements ClassDescriptor {
 
     @Override
     public MetadataInfoAnnotationDescriptor getMetadataInfoAnnotation() {
+
+        Optional<Annotation> metadataInfoAnnotation = getJavaAnnotationWithName("Lcom/telerik/metadata/MetadataInfo;");
+        if (metadataInfoAnnotation.isPresent()) {
+            return new MetadataInfoAnnotationInfo(dexFile, metadataInfoAnnotation.get());
+        }
+        return null;
+    }
+
+    private Optional<Annotation> getJavaAnnotationWithName(String name) {
         int annotationsOffset = classDef.getAnnotationsOffset();
         if (annotationsOffset == 0) {
-            return null;
+            return Optional.empty();
         }
 
         Dex dex = dexFile.getDex();
@@ -114,7 +170,7 @@ public class ClassInfo implements ClassDescriptor {
         int classDefIndex = dex.findClassDefIndexFromTypeIndex(classDef.getTypeIndex());
         int dexAnnotationDirectoryOffset = dex.annotationDirectoryOffsetFromClassDefIndex(classDefIndex);
         if (dexAnnotationDirectoryOffset == 0) {
-            return null;
+            return Optional.empty();
         }
         Dex.Section dirSection = dex.open(dexAnnotationDirectoryOffset);
         int classSetOffset = dirSection.readInt();
@@ -122,7 +178,7 @@ public class ClassInfo implements ClassDescriptor {
         int annotationSetOffset = classSetOffset;
 
         if (annotationSetOffset == 0) {
-            return null;
+            return Optional.empty();
         }
 
         Dex.Section annSetSection = dex.open(annotationSetOffset);
@@ -132,11 +188,22 @@ public class ClassInfo implements ClassDescriptor {
             Dex.Section annSection = dex.open(annotationOffset);
             Annotation annotation = annSection.readAnnotation();
             String annotationName = dex.strings().get(dex.typeIds().get(annotation.getTypeIndex()));
-            if (annotationName.equals("Lcom/telerik/metadata/MetadataInfo;")) {
-                return new MetadataInfoAnnotationInfo(dexFile, annotation);
+            if (annotationName.equals(name)) {
+                return Optional.of(annotation);
             }
         }
-        return null;
+
+        return Optional.empty();
+    }
+
+    @Override
+    public Optional<KotlinClassMetadataAnnotation> getKotlinClassMetadataAnnotation() {
+        Optional<Annotation> kotlinMetadataAnnotation = getJavaAnnotationWithName("Lkotlin/Metadata;");
+        if (kotlinMetadataAnnotation.isPresent()) {
+            KotlinClassMetadataAnnotation kotlinClassMetadataAnnotation = new KotlinDexClassMetadataAnnotation(kotlinMetadataAnnotation.get(), dexFile);
+            return Optional.of(kotlinClassMetadataAnnotation);
+        }
+        return Optional.empty();
     }
 
     @Override
@@ -145,7 +212,7 @@ public class ClassInfo implements ClassDescriptor {
         short[] ifaceIndexes = classDef.getInterfaces();
         if ((ifaceIndexes != null) && (ifaceIndexes.length > 0)) {
             List<String> typeNames = dexFile.getDex().typeNames();
-            for (short idx: ifaceIndexes) {
+            for (short idx : ifaceIndexes) {
                 String interfaceName = typeNames.get(idx);
                 names.add(interfaceName);
             }
@@ -168,7 +235,7 @@ public class ClassInfo implements ClassDescriptor {
         List<String> typeNames = dexFile.getDex().typeNames();
         int cdIdx = classDef.getTypeIndex();
         String typeName = typeNames.get(cdIdx);
-        String name = typeName.substring(1, typeName.length()-1).replace('/', '.');
+        String name = typeName.substring(1, typeName.length() - 1).replace('/', '.');
         return name;
     }
 
@@ -177,7 +244,7 @@ public class ClassInfo implements ClassDescriptor {
         List<String> typeNames = dexFile.getDex().typeNames();
         int idx = classDef.getSupertypeIndex();
         String superClassname = typeNames.get(idx);
-        String name = superClassname.substring(1, superClassname.length()-1).replace('/', '.');
+        String name = superClassname.substring(1, superClassname.length() - 1).replace('/', '.');
         return name;
     }
 }