+++ test-app/runtime/src/main/cpp/v8_inspector/src/inspector/protocol/Log.h
@@ -387,8 +387,8 @@ public:
     void entryAdded(std::unique_ptr<protocol::Log::LogEntry> entry);
 
     void flush();
-    void sendRawNotification(String);
-    void sendRawNotification(std::vector<uint8_t>);
+    void sendRawJSONNotification(String);
+    void sendRawCBORNotification(std::vector<uint8_t>);
 private:
     FrontendChannel* m_frontendChannel;
 };